Eliminate the redundant handling of index_select in get_splitfuse_attn_mask
When max_position_embeddings are large and max_seq_len is small, index_select will move a large amount of unused data, thereby increasing the utilization rate of the cpu, and the index_select operator executed on the cpu takes a very long time. this change can eliminate the redundant handling of index_select in get_splitfuse_attn_mask.

Signed-off-by: tt545571022 <tjl545571022@hotmail.com>